requirements.txt
absl-py==1.4.0
addict==2.4.0
anyio==3.6.2
astroid==2.9.3
attrs==23.1.0
basicsr==1.4.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
cmake==3.26.3
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
entrypoints==0.4
facexlib==0.3.0
fastjsonschema==2.16.3
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
filelock==3.12.0
filterpy==1.4.5
fonttools==4.39.3
future==0.18.3
gfpgan==1.3.8
google-auth==1.4.2
google-auth-oauthlib==1.0.0
grpcio==1.54.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
idna==2.8
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.6.4
ipython-genutils==0.2.0
isort==5.6.4
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
lazy_loader==0.2
lit==16.0.2
llvmlite==0.39.1
lmdb==1.4.1
logilab-common==1.8.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.6.1
mininet @ file:///mnt/c/Users/agust/OneDrive/Documentos/GitHub/Teleinformatica/mininet
mistune==2.0.5
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.3.0
mypy-extensions==0.4.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
ovs==2.17.3
packaging==23.1
pandocfilters==1.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.1
portpicker==1.2.0
proglog==0.1.10
prompt-toolkit==1.0.18
protobuf==4.22.3
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pydantic==1.10.7
pyflakes==2.4.0
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.12.2
pyparsing==2.4.7
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-telegram-bot==20.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.0.2
-e git+https://github.com/NathanFree/Real-ESRGAN-Easymode.git@6c18b8a7423b17229511aa803d1b01d11d97ee30#egg=realesrgan&subdirectory=esrgan
replicate==0.7.0
requests==2.21.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
SecretStorage==3.3.1
simplegeneric==0.8.1
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.1.0
soupsieve==2.4.1
ssh-import-id==5.11
sympy==1.11.1
systemd-python==234
tb-nightly==2.13.0a20230430
telegram==0.0.1
tensorboard-data-server==0.7.0
terminado==0.13.3
tifffile==2023.4.12
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==4.5.3
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.24.3
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.2
wrapt==1.13.3
yapf==0.33.0
zipp==1.0.0
